+++ src/models/selectedFormation.ts
@@ -1,0 +1,6 @@
+import Formation from "./formation";
+class SelectedFormation {
+    public formation: Formation;
+    public count: number;
+};
+export default SelectedFormation;