+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-# Oathmark-Army-Builder@1.2.1 (04.06.2021)
-- Add Halfling Army list
+# Oathmark-Army-Builder@1.3.0 (24.07.2021)
+- Implement Oathmark Bane of Kings
+    - Add stoney units
+    - Add chariots (for living armies)
+    - Add special formations
+- Divide unaligned Terrains into Unaligned (units), Expedition (campaign rewards), Support (Oathbreaker) and Formation